usbdux driver update from Bernd Porr:
This patch resolves problems with buggy VIA/SIS USB controllers. They tend to
loose IRQs in high speed mode. In the 2.6.4 kernel there's now a watchdog timer
which triggers an interrupt after 5ms. To make this safely work with the usbdux
driver the number of buffers has been increased to 10 for the high speed mode
and to 5 for the full speed mode.
There are also some minor changes: removed unnecessary defines and removed the
text that firmware upload through comedi is deprecated. In fact some distros
have fxload not installed so that comedi itself has (still) to do the job.