command support for pcl818 from Nils Juergens:
The streaming aquisition support has been dropped with the move from comedi_trig
to comedi_command. The attached patch ports the comedi_trig stuff over to
comedi_command. It is not heavily tested, but works for me with both pcl818 and
pcl818h cards (though i haven't tested the fifo code for the 818h).