patch from Ian Abbott <abbotti@mev.co.uk>, who says:
This patch fixes a harmless logical error in the _attach routines of the
amplc_pc236 and amplc_pc263 drivers that I noticed while looking at the
pci_for_each_dev stuff.  Also, now that the drivers ar using
pci_find_device instead of pci_for_each_dev, the patch offloads some of
the searching for a matching PCI device to the pci_find_device function.

The bug is just after the loop that finds the PCI device ID table entry
that matches the comedi board name.  The test for reaching the end of
the PCI device ID table without finding a match was incorrect.  I say
the bug is harmless because there would have to be another bug somewhere
else to trigger this bug.