patch from Ian Abbot <abbotti@mev.co.uk>:
I think it's probably a bad idea for my amplc_dio200 driver to call
comedi_event() while it's holding a spinlock.  The call to comedi_event
could call a kcomedilib client's callback function, which might do
something that requires the same spinlock, leading to a deadlock.

The attached patch moves the calls to comedi_event() to a safe point
after the spinlock is released.