patch from Ian Abbott <abbotti@mev.co.uk>:
The last change to das08_cs.c on Feb 5 breaks on 2.4, though the change
itself was not unreasonable (it saves a few bytes at least).  The
problem is that the change was incompatible with the version of struct
device_driver defined in comedi's linux/device.h compatibility header.
In the "real" version of struct device_driver, the "name" member is a
"char *", but in the compatibility version it's a "dev_info_t", which is
a "char[32]".

The attached patch sorts it out in the compatibility headers.