new driver from abbotti@mev.co.uk (Ian Abbott):
Here's a new driver for the Amplicon PCI224 and PCI234 analog output cards,
supporting streaming AO.  It seems to be working well except that the first scan
can misfire at the start of acquisition, and I can't find a way to work around
this hardware problem.  It shouldn't matter too much for waveform output.  I've
mentioned it in the caveats section of the driver documentation.