Sensoray 626 driver from gpalli@deis.unibo.it (Gianluca Palli):
This driver is based on the Sensoray 626 GPL Linux driver

The ai_poll function is not yet implemented

ao_cmd support is under development

The 48 dio channels are divided in to 3 subdev (2,3 and 4) with 16 channels
(each group of 16 channels is managed by a different hardware registers). These
can be grouped in one subdev with 48 bit mask. Channels 0-39 can be used as
TRIG_EXT arg for ai_cmd (cmd->start_src=TRIG_EXT; cmd->start_arg=20; subdev 3
channel 3 is used as external trigger)

Encoder/timer device has a lot of configuration option. Now only the encoder
configuration is avaiable. If ai_cmd scan_begin_src=TRIG_TIMER or
convert_src=TRIG_TIMER are selected, the timer channel 5 and 4 respectively are
used as trigger.

The code is filled with a lot of debugging messages visible compiling the code
with -D_DEBUG_ flag