Patch from terry1@beam.ltd.uk (Terry Barnaby):
This patch adds support to comedi for the NI PCI-6143 DAQ card.
It also adds support for the NI RTSI trigger bus used to synchronise multiple
cards to any ni_pcimio supported card.
The patch also requires a patch to the comedilib to support the RTSI.
More info on the patch is at: http://www.beam.org.uk/opensource/pci-6143/