Patch from calin@ajvar.org (Calin A. Culianu):
I modified the driver to more correctly reflect the physical connectors on the
PCMUIO board.  Basically rather than have 32-channel subdevices, I am making the
board use 24-channel subdevices.

The reason is simple: The board has 25-pin connectors.  Either two of them for
the 48-channel version or 4 of them for the 96-channel version.  Anyway, I made
this driver create 1 subdevice per connector.  It turns out in practice it is
more useful to have 4 subdevices with 24 channels each than it is to have 3
subdevices with 32 channels -- since you have to do funny math in your head to
figure out what channel 7 on subdevice 2 is.  By having a 1-to-1 association
between subdevice and physical connector, it's much easier to work with.