Applied patch from Ben Gardiner (blg at mast dot queensu dot ca):
It appears as though the changes were applied only to configure.ac.

In the current state, setting --disable-rtai will not avoid the check for the
presence of an rtai directory (line 16 in m4/rtai.m4) if the user has an
adeos/ipipe enabled kernel.

In the situation where the users has an adeos/ipipe enabled kernel, and a
non-standard rtai directory and does not wish to build comedi with rtai support
supplying --disable-rtai is not sufficient to complete the configuration script
without error. An additional --with-rtaidir=/non/standard/dir/ must be
specified. This seems counter-intuitive since the user has already supplied the
 --disable-rtai option.