New drivers from Jonathan Grierson <jd@renko.co.uk>:
Hi, I've recently been working with two digital
daq cards and have written drivers to get them working with Comedi.

The first is for the National Instruments PCI-6514. It could perhaps be
merged with the 6527 driver because I used this as a base but the 6514
is part of a series from NI which are all quite similar and not quite as
advanced. Let me know what you think. I'm afraid it needed to be done
quite quickly and so felt it would be easier to produce a simpler driver
rather than hack the existing one.

The second is for the Adlink PCI-7296. Just a pci wrapper for 4 8255
devices. Only basic IO enabled so far so no interrupts.