Patch from "Calin A. Culianu" <calin@ajvar.org>:
A pet peeve of mine is that the two driver that I wrote, the pcmda12 and
pcmuio driver, format incorrectly when displayed on the comedi.org
"Supported Hardware" web page.  This is because they are incorrectly
formatted in the driver's "Devices:" line.  This patch fixes this so that
the Devices: line is now correctly formatted.