comedi: Removal of <linux/smp_lock.h>
The <linux/smp_lock.h> header (used for the Big Kernel Lock) is removed
in kernel 2.6.39.  It was included by:

  comedi/comedi_compat32.c
  comedi/drivers/usbdux.c
  comedi/drivers/usbduxfast.c

The usbdux and usbduxfast drivers shouldn't need to include the header
as they don't grab the BKL, so remove the #include line for these
drivers.

The comedi_compat32 code only uses it when calling the ioctl file
operation, and that can only happen for kernels up to 2.6.35 (and even
then it wouldn't be called if the unlocked_ioctl file operation is
available!). Ifdef out the #include line.  Also include
<linux/version.h> explicitly to avoid compiler errors for 2.6.39 kernel.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>