amplc_pc263: Change DIO subdevice to DO.
This is a relay output card with no inputs.  Its subdevice is currently
configured as a DIO subdevice.  Change it to a DO subdevice and remove
the insn_config hook.

Also removed the deprecated SDF_RT flag.  Leave the SDF_READABLE flag.
Some drivers set this for DO subdevices and some don't.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>