ni_pcimio: fix minimum AO period for NI 625x and NI 628x
Follow patch submitted to the "staging" sources by Éric Piel, who writes:

The minimum period was set to 357 ns, while the divider for these boards
is 50 ns. This prevented to output at maximum speed as ni_ao_cmdtest()
would return 357 but would not accept it.

Not sure why it was set to 357 ns (this was done before the git history,
which starts 5 years ago). My guess is that it comes from reading the
specification stating a 2.8 MHz rate (~ 357 ns). The latest
specification states a 2.86 MHz rate (~ 350 ns), which makes a lot more
sense.

Tested on a pci-6251.

Thanks for the fix, Éric!

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>