Add linux/init.h compatibility header.
Add linux/init.h compatibility header, defining the __devinit,
__devinitdata, __devinitconst, __devexit, __devexitdata, __devexitconst
and __devexit_p(x) macros if they are not already defined.

This is needed for compatibility with Linux kernel 3.8 which removed the
macros.

We could just remove all uses of the macros from the code, but for now
do it this way.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>