rti800: fix straight-binary/2's-complement option for DAC channel 1
configuration option[8] is supposed to indicate whether the DAC channel
1 jumpers are set for straight binary or 2's complement coding, with
option[6] doing the same for DAC channel 0.  However, the code that
writes the DAC channels currently uses the option[6] value for both
channels.  Fix it to use the correct option for DAC channel 1.  Thanks
to H Hartley Sweeten for spotting this bug.

Also replace the dac0_range and dac1_range members in the private data
with array dac_range[2], and replace dac0_coding and dac1_coding members
with array dac_coding[2].

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>