comedi/comedi_fops.c: make COMEDI_CANCEL ioctl wake up reader/writer
If a task is doing a blocking read or write on a comedi subdevice, it
would be a good idea to wake it up if another task issues a
COMEDI_CANCEL ioctl on the file, otherwise it would just remain sleeping
even though the asynchronous command has been terminated.  Change
do_cancel_ioctl() in comedi_fops.c to do the wake-up after calling
do_cancel().

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>