ni_mio_common: handle AI FIFO dregs for M-series cards
ni_ai_handle_fifo_dregs() needs to handle M-series cards as it will be
called at the end of an AI command to extract any dregs of data from the
AI FIFO after DMA data has been dealt with.  It currently does a 16-bit
read of the ADC FIFO data register but should really do a 32-bit read
for M-series cards.  Also, for some board models the SDF_LSAMPL
subdevice flag is set, so it needs to handle the larger sample size for
these board models.

To handle the larger sample size, change devpriv->ai_fifo_buffer from an
array of sampl_t into a union containing an array s of sampl_t or an
array ls of lsampl_t (with half the number of elements).

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>