include/linux/comedidev.h, comedi_rt.h: conditionally include <linux/config.h>
The <linux/config.h> header is not needed for kernel version 2.6.15
onwards and was removed in 2.6.19 (or 2.6.18.x for Red Hat).  Our
wrapper header for <linux/config.h> copes with that and only includes
the kernel's <linux/config.h> for kernel versions prior to 2.6.15.

To avoid the comedi header files depending on the wrapper headers too
much, add conditional tests around the inclusion of <linux/config.h> in
<linux/comedidev.h> and <linux/comedi_rt.h>.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>