inc-wrap/linux/stddef.h: don't redefine bool
A previous commit from 2011-09-05 ("Deal with backported bool type for
RHEL's 2.6.18 kernel") conditionally defined the type 'bool' in the
<linux/types.h> wrapper, but left an incompatitible definition of the
type in the <linux/stddef.h> wrapper.  Remove the definition of 'bool'
from the <linux/stddef.h> wrapper but leave the 'true' and 'false' enum
constants defined there.

Note that the <linux/types.h> wrapper uses the type '_Bool' to define
the 'bool' type.  This assumes the GCC compiler is recent enough to
define the built-in type '_Bool'.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>