m4/rtai.m4: allow --with-rtaidir to specify RTAI build directory
The --with-rtaidir=DIR option currently only works if DIR is the RTAI
installation directory (defaulting to "/usr/realtime").  Try and make it
work for an RTAI build directory or a combined RTAI build and source
directory as well.  It probably won't work unless RTAI has been built in
the specified directory.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>