ni_atmio16d: handle INSN_CONFIG properly for DIO subdevice
Change atmio16d_dio_insn_config() to treat data[0] as a configuration
instrunction number instead of treating the data[] as an array of
booleans.  The 'for' loop is pointless, so get rid of it.  Handle
INSN_CONFIG_DIO_INPUT, INSN_CONFIG_DIO_OUTPUT and INSN_CONFIG_DIO_QUERY.
As long as old application code was only setting data[0] to 0 or 1 and
setting insn->n to 1, that should still work the same.  More bizarre
data[] settings would likely cause the comedi core to return an error
anyway.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>