ni_65xx: (bug fix) confine insn_bits operation to single subdevice
The insn_bits handler ni_65xx_dio_insn_bits() has a loop that processes
up 5 ports consisting of 8 channels each (it wants to handle at most 32
channels, but the first channel might not be aligned to a port
boundary).  It breaks out of the loop early if the current port number
is beyond the last port on the card, and also breaks out if it has
handled at least 32 channels, but it doesn't break out if the port does
not belong to the subdevice.  That's a bug.

Rearrange the loop to restrict it to the ports belonging to the
subdevice.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>