cb_pcidio: Cope with change of PCI interface chip
At least for the PCI-DIO48H, and possibly others, the hardware design
seems to have changed from using an AMCC S5933 (I think) PCI interface
chip with the user registers in PCI BAR 1, to using a PLX PCI9050 or
PCI9052 with the user registers in PCI BAR 2.

Since the driver currently uses PCI BAR 1 for PCI-DIO48H, it won't work
for newer versions of the card.  Also, since the driver uses PCI BAR 2
for the other supported cards, it might not work for older versions of
those cards.

Use the PCI BAR 2 region if it is non-zero length, falling back to using
the PCI BAR 1 region.  That should work for old and new cards.

If the driver ever supports interrupts in the future, the difference in
PCI interface chips will have to be taken into consideration!

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>