usbduxsigma: use get_unaligned() for unaligned reading of insnBuffer
For simple AI read received data, the URB data buffer pointer is a
'int8_t *', the first byte contains a copy of the command byte and the
next 4 bytes contain a big-endian value containing status and data
value.  Currently this is accessed by casting a pointer to this 4-byte
value to a 'int32_t *' and dereferencing.  That works for x86, but may
not work as expected for other architectures.  Use the get_unaligned()
macro to dereference the pointer value for portability.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>