dt282x: Fix reporting of error for command on AO subdevice
In the interrupt handler dt282x_interrupt(), if the DACSR register
indicated an error had occurred on the AO subdevice, the COMEDI_CB_ERROR
flag was being set in the AI subdevice instead of the AO subdevice.  Fix
it.  Also call comedi_event() on the AO subdevice as well as the AI
subdevice.

Spotted by H Hartley Sweeten.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>