dt282x: correct options to make D/A conversions use 2's complement
The two 'attach' options to set whether the two AO channels use 2's
complement encoding or not (options 2 and 3) are checked incorrectly
when writing to the A0 channels.  I.e. the option for channel 1 is used
when writing to channel 0 and vice versa.  Fix the code to use the
correct options.

Spotted by H Hartley Sweeten in the "staging" drivers.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>