cb_pcidas64: fix AO stop count
'priv(dev)->ao_count' is set to the total number of samples to be
written (if 'async->cmd.stop_src' is 'TRIG_COUNT') but
'load_ao_dma_buffer()' currently treats it as a number of bytes, so the
async command will terminate when only half the total number of samples
have been written.  Fix it.

Thanks to Hartley Sweeten for spotting the bug in the "staging" drivers.

There is another problem when the number of scans multiplied by the
number of channels per scan overflows a 32-bit unsigned int.  In that
case, the total number of samples will be less than it should be.  That
is not addressed by this patch.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>