das16: fix setting of pacer when scan_begin_src == TRIG_TIMER
When scan_begin_src == TRIG_TIMER (and convert_src == TRIG_NOW), the
pacer is not set up correctly - it uses convert_arg instead of
scan_begin_arg.  Also, convert_arg would be 0 in that case, which would
be interpreted as setting the maximum possible period that the timers
can be set to (4.29e9 times the pacer clock period).  Fix it by setting
the pacer period to convert_arg if convert_src == TRIG_TIMER, or to
scan_begin_arg if scan_begin_arg == TRIG_TIMER.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>