adl_pci6208: Treat all boards as PCI-6216V
All boards supported by this driver have the same PCI device ID.
Previously, detected boards were treated as a PCI-6208V, although the
board name was set to "pci6208a" for some reason!  This means that if a
PCI-6216V is plugged in, only the first 8 AO channels are available.

Change it to treat all detected boards as a PCI-6216V with board name
"pci6216v".  The AO subdevice will now have 16 channels.  It should
still work with a PCI-6208V plugged in, but channels 8 to 15 will be
"phantom" outputs.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>