adl_pci6208: Add DI and DO subdevices
In addition to the analog outputs, the boards supported by this driver
have 4 digital inputs and 4 digital outputs, but the digital inputs and
outputs are not currently supported.  The driver allocates 2 subdevices,
but subdevice 1 is unused.

Change it to allocate 3 subdevices, assigned as follows:

Subdevice 0: analog outputs
Subdevice 1: digital inputs
Subdevice 2: digital outputs

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>