ni_65xx: change async command to write DIO and DI channels to buffer
The interrupt handler for asynchronous commands on the "change of state
notification" subdevice 3 currently just writes a sample value 0 to the
buffer.  If the board has DIO or DI channels, it would be useful to
write the input data to the buffer.

Set the SDF_PACKED flag on subdevice 3.  If the board has DIO or DI
channels, set subdevice 3 to have that number of channels, otherwise set
it to 1.  If the board has DIO or DI channels, change the interrupt
handler to write all the DIO or DI channels to the comedi buffer in
packed format (up to 16 channels per sampl_t), otherwise just write a 0
to the buffer.

Inspired by a patch by Maxime Chevallier <chevalliersebmax at gmail dot
com>.  His patch only considered DI channels and wrote the channel data
to the buffer in the "wrong" order.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>