comedi.h: Force some constants to be unsigned to help SWIG
The CR_FLAGS_MASK, CR_INVERT, and TRIG_ANY macro constants all have bit
31 set.  Force them to be unsigned to stop SWIG turning them into
negative integers.  This follows a patch to comedi.h within COMEDILIB by
Éric Piel, and keeps the two versions of the include file identical.

There is another problem constant NI_GPCT_INVERT_CLOCK_SRC_BIT with the
value 0x80000000, but that is defined as an enumerated constant and by C
rules is treated as being of type 'int' so is negative.  COMEDILIB's
comedi.i SWIG interface file redefines that constant to be unsigned as a
special case.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>