adl_pci7250: New driver for ADLINK PCI-7250
Adds support for the ADLINUX PCI-7250 8-CH relay output plus 8-CH
isolated digital input card as well as the LPCI-7250 and LPCIe-7250.

The driver has not been tested yet.

Note that the driver actually sets the number of outputs and inputs to
32 channels each, as it allows up to three PCI-7251 extension cards to
be fitted, providing an additional 8 channels of inputs and 8 channels
of relay outputs each.  The LPCI-7250 and LPCIe-7250 do not actually
support the PCI-7251 extension cards, but the driver assumes they are
fitted anyway!  Non-fitted extension cards shouldn't do any harm, but
the extra channels won't do anything.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>