cb_pcimdas: only read the SE/DIFF switch during attach
The board has either 8 differential channels or 16 single-ended channels
according to the setting of switch S1.  The state of the switch can be
read through a register.  Currently, this is done in the insn_read
handler for the AI subdevice.  It has to return an error when reading an
out-of-range channel number in differential mode.

Change it to read the state of the switch once during "attach", and set
up the number of channels and subdevice analog reference flags
accordingly.  Remove the test for out-of-range channels from the AI
insn_read handler. We assume the switch state will not change during
operation.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>