cb_pcimdas: support programmable gain for AI subdevice
The hardware has a programmable gain register allowing a choice of 4
bipolar or 4 unipolar ranges for analog input.  The choice between
bipolar or unipolar is selected by a switch whose state can be read in a
register.

Currently, the driver ignores the bipolar/unipolar switch and always
programs the gain to the widest range (+/- 10V bipolar, or 0-10V
unipolar).  It declares the range of the AI subdevice to be "unknown".

Change it to support programmable gain on the AI subdevice.  Read the
state of the Bip/Uni switch during "attach" and choose set the subdevice
to use one of two range tables (unipolar or bipolar) accordingly.  Each
range table supports 4 ranges.  Bipolar ranges: +/-10V, +/-5V, +/-2.5V,
+/-1.25V.  Unipolar ranges: 0-10V, 0-5V, 0-2.5V, 0-1.25V.

Thanks to Hartley Sweeten for implementing this in the "staging"
drivers.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>