das1800: bug fix bipolar munging for AI commands
When a unipolar range is used, some samples read by the AI command will
be incorrectly "munged" as bipolar data due to a bug in
`das1800_handle_fifo_not_empty()`.  Other samples will be handled
correctly by `das1800_handle_fifo_half_full()` or by
`das1800_handle_dma()`.  Typically, the data in the buffer will
alternate by correctly munged and incorrectly munged data.

The bug in `das1800_handle_fifo_not_empty()` is that the function tests
whether a bipolar or unipolar range is being used, but ignores the test
and munges the sample data as bipolar regardless.  Fix the bug by only
munging the sample data if a bipolar range is in use.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>