adv_pci1710: Fix bug setting D/A range for PCI-1716
For PCI-1716, the D/A reference control register is only being updated
when the new range differs from the previous value, even if the previous
value was for the other channel.  Fix it by always writing the D/A
reference control register.

Also, always treat the D/A reference control register as a 16-bit
register (even for card types where only the bottom 8-bits are defined)
and increase the width of its shadow in the 'da_ranges' member of the
private data accordingly.  Some of the existing code already writes it
as a 16-bit register, so this should be safe.

Also, set the D/A range of the PCI-1716 to the 0 - 5V unipolar range
when the reset function is called during "attach" and "detach" to make
it the same as the other supported card types.  Since 0 is written to
the D/A data registers during this reset, this should result in the
output voltage being set to 0V.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>