adv_pci1710: Support external D/A ranges
PCI-1710, PCI-1710HG, PCI-1711, and PCI-1716 support an external source
for the D/A range.  Support this in the range tables.

The PCI-1716 supports both bipolar and unipolar ranges, but the comedi
range index corresponding to the unipolar external range (5) does not
match the hardware range code (6), so handle that as a special case.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>