Modify LINUXINCLUDE instead of CC
We need the Comedi header directories (in particular the 'inc-wrap'
directory) to be searched before the Linux kernel header directories.
This was done by modifying the 'CC' variable, but that broke when
building Comedi for the 4.15 kernel.  Modify the 'LINUXINCLUDE' variable
instead.  The extra options are exported from "comedi/Makefile" to
"comedi/Kbuild" via the environment variable 'COMEDI_PREPEND_CFLAGS'.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>