Create <linux/timer.h> compatibility header
So far, this <linux/timer.h> header only defines the timer_setup()
function and the from_timer() macro, which both appeared in Linux kernel
4.14.  We will need to migrate the Comedi timer stuff to use these
functions in order to build for kernel version 4.15 and beyond.  We
ignore the final 'flags' parameter of timer_setup(), so Comedi code
should set it to 0.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>