cb_pcidas64: Fix inverted test in external_ai_queue_in_use()
The initial test in external_ai_queue_in_use() for the "read" subdevice
being "busy" is the inverse of what it should be.  The function should
return 0 if the "read" subdevice is not busy because the external AI
queue cannot be in use in that case.

The original, inverted test could also result in a null pointer
dereference in the subsequent call to use_internal_queue_6xxx() when a
previous AI command has a channel list length greater than 1.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>