Add 'comedi_poll_t' typedef
The 4.16 kernel changed 'poll' file operations to use the EPOLL*
constants (e.g. EPOLLIN) instead of the POLL* constants (e.g. POLLIN).
It also defined the '__poll_t' type to hold values of this type and
changed the return type of the 'poll' file operations to '__poll_t'.

In an earlier commit, we defined the COMEDI_EPOLL* macros to expand to
EPOLL* or POLL* depending on the kernel version, but we didn't do
anything about the '__poll_t' type.  Add a typedef to define
'comedi_poll_t' from '__poll_t' for newer kernels, or from 'unsigned
int' for older kernels, and change the comedi_poll() function in
"comedi_fops.c" to use this type.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>