quatech_daqp_cs: Replace strncpy call with strlcpy to avoid GCC warning
GCC 8 warns about strncpy() calls where the size equals the destination
buffer size (if that can be determined), as that could leave the
destination without a null terminator.

Replace the strncpy() call with strlcpy().  This ensures the destination
is null-terminated.

If the source string is 32 or more characters long, the 32nd character
is no longer copied, but Comedi only uses 20 character board and driver
names anyway.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>