m4/as-linux.m4: remove AS_LINUX_EXTRAVERSION and --with-extraversion
The AS_LINUX_EXTRAVERSION Autoconf function sets the LINUX_EXTRAVERSION
variable from a passed-in `uname -r`-style kernel version string.  It
also allows this to be overridden with a '--with-extraversion='
command-line option.  However, nothing uses the LINUX_EXTRAVERSION
variable, so get rid of AS_LINUX_EXTRAVERSION.  (This also results in
the '--with-extraversion=' configure command-line option no longer
listed by the '--help' option.)

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>