m4/as-linux.m4: Fix AS_LINUX_SRC_DIR for 4.20 kernel
The tests in the AS_LINUX_SRC_DIR Autoconf function that check for
separate Linux source and build directory fail for kernel version 4.20
due to changes in the generated Makefile for the separate kernel build
directory.  Add an extra test to detect the kernel source directory
referred to by a kernel 4.20 build directory Makefile.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>