inc-wrap/ktime.h: New wrapper with do_gettimeofday()
Add a kernel compatibility wrapper for <linux/ktime.h> which is only
included for kernel version 2.6.16 or later.

The do_gettimeofday() function was removed in kernel version 4.20, but
our Comedi kernel code still uses it.  Add an inline version here for
kernel version 4.20 onwards until we get around to changing the code
that calls it.  (Or maybe we'll just leave it as-is!)

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>