dmm32at: Fix for loop in dmm32at_ao_winsn()
dmm32at_ao_winsn() has an outer 'for' loop to iterate through the list
of samples to be written to the device, and an inner 'for' loop to wait
for the DAC to become non-busy.  Unfortunately, they are using the same
loop variable 'i'.  Introduce a new loop variable 'j' for the inner
'for' loop.

Thanks to Paolo Mantegazza for providing the fix.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>