adv_pci1724: Support synchronized output mode
Use an INSN_CONFIG instruction of data length 2 and data[0] set to
INSN_CONFIG_ARM to enable synchronized output mode, and optionally
trigger synchronized output.  If data[1] is 0, synchronized output mode
is only enabled.  Setting data[1] to a non-zero value triggers
synchronized output.

Use an INSN_CONFIG configuration instruction of data length 1 and
data[0] set to INSN_CONFIG_DISARM to disable synchronized output mode.

The above instructions are only supported for subdev 0 and the chanspec
should normally be set to 0 (the channel number needs to be valid, but
is not used).

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>