inc-wrap/linux/time.h: Add struct timeval for kernel version 5.6+
struct timeval was removed in Linux kernel version 5.6, but some of our
Comedi code still uses it.

For now, declare struct timeval if building for kernel version 5.6 or
later, but the code that uses it ought to be updated to use something
else, but that will also compile on earlier kernels we wish to keep on
supporting.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>