comedi_fops: Big change to allow per-open file read and write subdevices
Get rid of 'struct comedi_device_file_info' and
'comedi_file_info_table[]'.  The table is replaced by
'comedi_board_minor_table[]' holding pointers to 'comedi_device' for
regular comedi device files and with 'comedi_subdevice_minor_table[]'
for comedi subdevice files.  The class device's drvdata is now set to
the minor device number.

When a comedi file is opened, allocate a new 'struct comedi_file' as the
device private data.  This records the 'comedi_device', the read
subdevice (if any) and the write subdevice (if any) associated with the
open file description.  This will allow the read and write subdevice to
be changed dynamically on a per-open file description using ioctl
commands (not yet implemented, but they will be the same as those
supported by the "staging" version of COMEDI included in the Linux
kernel).

Add a 'detach_count' member to the 'comedi_device' type, incremented
each time the comedi device is detached.  This is compared with the
'detach_count' in 'struct comedi_file' to check the validity of previous
per-file changes to the read and write subdevices.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>

squash! comedi_fops: Big change to allow per-open file read and write subdevices